requirements.web.txt
fastapi[all]==0.115.8
uvicorn[standard]==0.22.0
sqlalchemy==2.0.40
asyncpg==0.30.0
httpx==0.28.1
pydantic>=2.7.0,<3.0
pydantic-settings==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
feedparser==6.0.10
apscheduler==3.10.1 # 定時配信型にしたい場合は必要
line-bot-sdk==3.17.1
google-genai==1.11.0
janome==0.5.0
gensim==4.3.3
python-dateutil==2.9.0.post0
deepl
